Fix specs according to bower removal refactoring
